+++ src/supplier/dto/create-supplier.dto.ts
@@ -11,29 +11,3 @@ export class CreateSupplierDto {
   @MinLength(10)
   nomor_telepon: string;
 }
-
-export class PaginationSupplierDto {
-  @IsInt()
-  @Min(1)
-  readonly page: number;
-
-  @IsInt()
-  @Min(1)
-  readonly limit: number;
-
-  @IsOptional()
-  @IsIn(['ASC', 'DESC'])
-  readonly orderNama?: 'ASC' | 'DESC';
-
-  @IsOptional()
-  @IsIn(['ASC', 'DESC'])
-  readonly orderAlamat?: 'ASC' | 'DESC';
-
-  @IsOptional()
-  @IsIn(['ASC', 'DESC'])
-  readonly orderNomorTelepon?: 'ASC' | 'DESC';
-
-  @IsOptional()
-  @Matches(/^[A-Za-z0-9\s]*$/)
-  readonly keywords?: RegExp;
-}
