+++ src/supplier/supplier.controller.ts
@@ -14,10 +14,8 @@ import {
 } from '@nestjs/common';
 import { Supplier } from './entities/supplier.entity';
 import { SupplierService } from './supplier.service';
-import {
-  CreateSupplierDto,
-  PaginationSupplierDto,
-} from './dto/create-supplier.dto';
+import { CreateSupplierDto } from './dto/create-supplier.dto';
+import { PaginationSupplierDto } from './dto/pagination-supplier.dto';
 import { UpdateSupplierDto } from './dto/update-supplier.dto';
 import { JwtAuthGuard } from 'src/auth/auth.guard';
 import { SupplierResponse } from './types/supplier.response.type';
